+++ CHANGELOG.md
@@ -1,3 +1,94 @@
+## v0.34.0
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- **cloudflare**: Add preview IDs for KV, D1, and R2 in wrangler.json &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/413 [<samp>(4a59d)</samp>](https://github.com/sam-goodwin/alchemy/commit/4a59d3a3)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.33.1...v0.34.0)
+
+---
+
+## v0.33.1
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- **cloudflare**: Use dispatch namespace asset upload for WFP &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/412 [<samp>(bf6b5)</samp>](https://github.com/sam-goodwin/alchemy/commit/bf6b5b80)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.33.0...v0.33.1)
+
+---
+
+## v0.33.0
+
+### &nbsp;&nbsp;&nbsp;🚀 Features
+
+- **stripe**: Add Full Tier Support for Stripe Prices &nbsp;-&nbsp; by **Nick Balestra-Foster** in https://github.com/sam-goodwin/alchemy/issues/406 [<samp>(7691c)</samp>](https://github.com/sam-goodwin/alchemy/commit/7691c2ce)
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- **cloudflare**: Configure Websites to support SSR by default &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/409 [<samp>(72521)</samp>](https://github.com/sam-goodwin/alchemy/commit/72521e8f)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.32.1...v0.33.0)
+
+---
+
+## v0.32.1
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- **cloudflare**: Fix SSR for astro by setting not_found_hanlding=none &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/407 [<samp>(7750f)</samp>](https://github.com/sam-goodwin/alchemy/commit/7750f024)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.32.0...v0.32.1)
+
+---
+
+## v0.32.0
+
+### &nbsp;&nbsp;&nbsp;🚀 Features
+
+- **cloudflare**: Add Worker version/preview support with labels &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/351 [<samp>(615fc)</samp>](https://github.com/sam-goodwin/alchemy/commit/615fc9d1)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.31.0...v0.32.0)
+
+---
+
+## v0.31.0
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- Bug in create vitejs &nbsp;-&nbsp; by **Sam Goodwin** [<samp>(4efe2)</samp>](https://github.com/sam-goodwin/alchemy/commit/4efe232e)
+- Move required and internal peerDeps to deps &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/402 [<samp>(31d92)</samp>](https://github.com/sam-goodwin/alchemy/commit/31d924d2)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.30.1...v0.31.0)
+
+---
+
+## v0.30.1
+
+### &nbsp;&nbsp;&nbsp;🚀 Features
+
+- **cloudflare**:
+  - Auto-create default secrets store &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/395 [<samp>(3251d)</samp>](https://github.com/sam-goodwin/alchemy/commit/3251d1e1)
+  - Add SecretKey binding support for Cloudflare Workers &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/385 [<samp>(08278)</samp>](https://github.com/sam-goodwin/alchemy/commit/082785a1)
+
+### &nbsp;&nbsp;&nbsp;🐞 Bug Fixes
+
+- **cloudflare**: Secret binding type &nbsp;-&nbsp; by **Tyler van Hensbergen** in https://github.com/sam-goodwin/alchemy/issues/398 [<samp>(724d9)</samp>](https://github.com/sam-goodwin/alchemy/commit/724d9c26)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.30.0...v0.30.1)
+
+---
+
+## v0.30.0
+
+### &nbsp;&nbsp;&nbsp;🚀 Features
+
+- Implement alchemy create CLI &nbsp;-&nbsp; by **Sam Goodwin** in https://github.com/sam-goodwin/alchemy/issues/376 [<samp>(85a8e)</samp>](https://github.com/sam-goodwin/alchemy/commit/85a8e2f8)
+
+##### &nbsp;&nbsp;&nbsp;&nbsp;[View changes on GitHub](https://github.com/sam-goodwin/alchemy/compare/v0.29.3...v0.30.0)
+
+---
+
 ## v0.29.3
 
 ### &nbsp;&nbsp;&nbsp;🚀 Features